Fix sysFeatures amd rapl not initializing at all
This commit also distributes header and body code file to be more in
line with the Intel RAPL code. AMD RAPL code is not yet checked for
semantic bugs.
